+++ app/src/main/java/com/wearerommies/roomie/presentation/ui/mood/navigation/MoodNavigation.kt
@@ -1,4 +1,4 @@
-package com.wearerommies.roomie.presentation.ui
+package com.wearerommies.roomie.presentation.ui.mood.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
@@ -6,6 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.wearerommies.roomie.presentation.navigator.route.Route
+import com.wearerommies.roomie.presentation.ui.mood.MoodRoute
 
 fun NavController.navigateToMood(navOptions: NavOptions? = null) {
     navigate(
@@ -19,9 +20,9 @@ fun NavGraphBuilder.moodNavGraph(
     navigateUp: () -> Unit
 ) {
     composable<Route.Mood> {
-//        MoodRoute(
-//            paddingValues = paddingValues,
-//            navigateUp = navigateUp
-//        )
+        MoodRoute(
+            paddingValues = paddingValues,
+            navigateUp = navigateUp
+        )
     }
 }
